making sure current user is the owner before allowing edit or delete

+++ clktc/links/views.py
@@ -1,6 +1,6 @@
 # Create your views here.
 from django.contrib.auth.decorators import login_required
-from django.http import  HttpResponseNotAllowed
+from django.http import  HttpResponseNotAllowed, HttpResponseNotFound
 from django.shortcuts import render_to_response, redirect, get_object_or_404
 from django.template.context import RequestContext
 from clktc.links.forms import AddLinkForm, EditLinkForm
@@ -33,6 +33,8 @@ def add_link(request):
 @login_required
 def edit_link(request, link_id):
     link = get_object_or_404(Link, pk=link_id)
+    if link.user != request.user:
+            return HttpResponseNotFound()
     form_cls = EditLinkForm
     if request.method == "POST":
         form = form_cls(request.POST, instance=link)
@@ -48,6 +50,8 @@ def edit_link(request, link_id):
 def delete_link(request, link_id):
     if request.method == "POST":
         link = get_object_or_404(Link, pk=link_id)
+        if link.user != request.user:
+            return HttpResponseNotFound()
         link.delete()
         return redirect(get_all_links)
     else:
